# exchange-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exchange Rate Application",
    "description": "This is an exchange rate application where users can sign up for an account, add transactions, see their previous transactions, and get the current exchange rate calculated according to the past 3 days. Users can also check the conversion rate and amount for buying or selling USD and LBP.",
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/"
    }
  ],
...
  "paths": {
...
    "/transaction": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "summary": "Create a new transaction",
        "operationId": "createTransaction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "usd_amount",
                  "lbp_amount",
                  "usd_to_lbp"
                ],
                "properties": {
                  "usd_amount": {
                    "type": "number",
                    "format": "float",
                    "description": "Amount in USD"
                  },
                  "lbp_amount": {
                    "type": "number",
                    "format": "float",
                    "description": "Amount in LBP"
                  },
                  "usd_to_lbp": {
                    "type": "boolean",
                    "description": "Conversion direction, USD to LBP if true"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Validation error for input data"
          },
          "403": {
            "description": "Authentication error, token invalid or expired"
          }
        },
        "security": [
          {
            "None": []
          },
          {
            "JWTToken": []
          }
        ]
      },
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Retrieve all transactions for the authenticated user",
        "operationId": "getTransactions",
        "responses": {
          "200": {
            "description": "A list of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Authentication error, token invalid or expired"
          }
        },
        "security": [
          {
            "JWTToken": []
          }
        ]
      }
    },
...
    "/user": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create new user",
        "description": "A new user can create an account for themselves.",
        "security": [],
        "operationId": "createUser",
        "requestBody": {
          "$ref": "#/components/requestBodies/UserRequestBody"
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Input empty, username not valid, or password is weak.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Weak password"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User already exists"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user from token",
        "description": "Retrieves user information based on the provided JWT token.",
        "responses": {
          "200": {
            "description": "User retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "498": {
            "description": "Invalid or expired token."
          },
          "499": {
            "description": "Token not provided."
          }
        },
        "security": [
          {
            "JWTToken": []
          }
        ]
      }
    },
...
    "/wallet": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "Retrieve wallet values",
        "description": "Retrieves the wallet values for the user authenticated by the token provided in the request headers.",
        "responses": {
          "200": {
            "description": "Wallet values retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wallet"
                }
              }
            }
          },
          "400": {
            "description": "Invalid token or no wallet associated with this account."
          }
        },
        "security": [
          {
            "JWTToken": []
          }
        ]
      }
    },
    "/wallet/updateWallet": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Update wallet values",
        "description": "Updates the wallet balances for the specified user ID with the provided USD and LBP values to add.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "usd_to_add",
                  "lbp_to_add",
                  "user_id"
                ],
                "properties": {
                  "usd_to_add": {
                    "type": "number",
                    "example": -100,
                    "description": "USD amount to add to the wallet."
                  },
                  "lbp_to_add": {
                    "type": "number",
                    "example": 200000,
                    "description": "LBP amount to add to the wallet."
                  },
                  "user_id": {
                    "type": "integer",
                    "example": 1,
                    "description": "User ID of the wallet owner."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Wallet updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wallet"
                }
              }
            }
          },
          "400": {
            "description": "Missing values, incorrect input names, or no wallet found for the given user ID."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "The transaction identifier"
          },
          "usd_amount": {
            "type": "number",
            "format": "float",
            "description": "Amount in USD"
          },
          "lbp_amount": {
            "type": "number",
            "format": "float",
            "description": "Amount in LBP"
          },
          "usd_to_lbp": {
            "type": "boolean",
            "description": "Conversion direction, USD to LBP if true"
          },
          "added_date": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the transaction was added"
          },
          "user_id": {
            "type": "integer",
            "description": "The identifier of the user who made the transaction"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "user_name": {
            "type": "string",
            "example": "omar1212"
          }
        }
      },
...
      "Wallet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "description": "Unique identifier of the wallet."
          },
          "user_id": {
            "type": "integer",
            "example": 1,
            "description": "Unique identifier of the user who owns the wallet."
          },
          "usd_amount": {
            "type": "number",
            "example": 500,
            "description": "Amount of USD in the wallet."
          },
          "lbp_amount": {
            "type": "number",
            "example": 10000000,
            "description": "Amount of LBP in the wallet."
          }
        }
      }
    },
    "requestBodies": {
...
      "UserRequestBody": {
        "description": "User data for registration",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "user_name": {
                  "type": "string",
                  "example": "omar1212"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "example": "Pass_1212"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "JWTToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "JSON Web Token (JWT) allow users to link transactions to themselves.\nAdd the JWT token in the format `Bearer <token>`.\n"
      },
      "None": {
        "type": "http",
        "scheme": "none",
        "description": "Authentication is not needed always.\n"
      }
    }
  }
}